feat: Update baanks transactions regexp
We were currently looking for "cpam", "c.p.a.m" or "caisse primaire".
But sometimes the transaction label contains "CPCAM". We have the case
for a user from Lyon. This PR adds an optional "c" on the regex, so we
also look for "cpcam" and "C.P.C.A.M"